requirements.txt

Flask~=2.2.2
google~=3.0.0
oauthlib~=3.1.1
requests~=2.28.0
bootstrap-flask~=2.1.0
jinja2~=3.1.2
markupsafe~=2.1.1
flask-restx~=0.5.1
Flask-Minify~=0.38
Flask-WTF~=1.0.1
Flask-SQLAlchemy~=2.5.0
PyMySQL~=1.0.1
python-dotenv~=0.21.0
SQLAlchemy~=1.4.41
flask-marshmallow==0.14.0
marshmallow~=3.17.1
gunicorn~=20.1.0
CacheControl~=0.12.11
flask_migrate~=3.1.0
# flask-assets~=2.0
Werkzeug==2.2.2
google-auth~=2.8.0
google-api-python-client~=2.60.0
google-auth-httplib2~=0.1.0
google-auth-oauthlib~=0.5.2

# mysql-python~=1.2.5
# mysqlclient~=2.1.1
mysqlclient~=2.1.1
WTForms~=3.0.1
PyYAML~=6.0
html2text~=2020.1.16